Ignore stale image download completion events;
When using `BFRBackLoadedImageSource`, if the hi-res image
was downloading slowly, it was possible that the original
view controller was already dismissed and another controller
would receive the global `NOTE_HI_RES_IMG_DOWNLOADED` for
the now-stale image request.

The fix attempted here is to make the NSNotificationCenter
event names unique by including the URL in the event name.